Added check for null object in getSoundData
Summary: Fixes T159, checks if object is null so pictogram without sound can be saved

Test Plan:
No unit test are added in this fix since the StoragePictogram object uses a
constructor which calls a constructor on a database helper and therefore cannot the object be mocked.
Furtheremore is the fixed function private.

Reviewers: kstick13

Reviewed By: kstick13

Maniphest Tasks: T159

Differential Revision: http://web.giraf.cs.aau.dk/D4